# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Plataforma de Auditoria Fiscal Preventiva (CFDI Intelligence)
  version: 1.0.0
  description: |
    API REST para la plataforma de deteccion automatica de riesgos fiscales.
    Permite gestionar empresas, ingestar CFDIs y estados de cuenta bancarios,
    conciliar movimientos, detectar riesgos fiscales, calcular scoring de
    cumplimiento y descargar CFDIs directamente del SAT via FIEL (e.firma).

    **Autenticacion**: Todos los endpoints (excepto /auth/register, /auth/login
    y /health) requieren un Bearer Token JWT obtenido al autenticarse.

    **Vencimiento del token**: 8 horas desde su emision.
servers:
  - url: https://api.cfdi-intelligence.com
    description: Produccion
  - url: http://localhost:8000
    description: Desarrollo local
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/empresas/{empresa_id}/conciliaciones:
    get:
      operationId: listarConciliaciones
      tags: [Conciliacion]
      summary: Resumen de conciliacion bancaria
      description: |
        Retorna un resumen de los tipos de match en la conciliacion bancaria.
        El porcentaje de conciliacion incluye matches exactos y parciales.
      parameters:
        - $ref: '#/components/parameters/EmpresaId'
        - name: periodo
          in: query
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
          description: Filtrar por periodo YYYY-MM
          example: "2024-01"
      responses:
        '200':
          description: Resumen de conciliacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumenConciliacion'
              example:
                total: 45
                exacto: 30
                parcial: 8
                sin_cfdi: 5
                sin_movimiento: 2
                pct_conciliado: 84.4
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/empresas/{empresa_id}/conciliaciones/accionables:
    get:
      operationId: conciliacionesAccionables
      tags: [Conciliacion]
      summary: Conciliaciones accionables
      description: |
        Retorna los pares sin_cfdi y parciales con contexto completo del
        movimiento bancario. Util para mostrar al usuario que movimientos
        necesitan accion. Limitado a 100 registros por consulta.
      parameters:
        - $ref: '#/components/parameters/EmpresaId'
        - name: periodo
          in: query
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
          description: Filtrar por periodo YYYY-MM
          example: "2024-01"
      responses:
        '200':
          description: Pares accionables de conciliacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConciliacionesAccionablesResponse'
              example:
                total: 5
                pares:
                  - id: "c1c2c3c4-0001-0001-0001-000000000001"
                    tipo_match: sin_cfdi
                    monto_movimiento: 15000.00
                    monto_cfdi: null
                    diferencia: 15000.00
                    porcentaje_match: 0.0
                    periodo: "2024-01"
                    movimiento_id: "m1m2m3m4-0001-0001-0001-000000000001"
                    mov_fecha: "2024-01-15"
                    concepto: "DEPOSITO CLIENTE XYZ"
                    mov_monto: 15000.00
                    mov_tipo: deposito
                    rfc_detectado: "XAXX010101000"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  # ============================================================
  # CIERRE
  # ============================================================
  /api/v1/empresas/{empresa_id}/periodos:
    get:
      operationId: listarPeriodos
      tags: [Cierre]
      summary: Listar periodos con datos
      description: |
        Lista los ultimos 24 periodos (YYYY-MM) que tienen datos cargados
        para la empresa (CFDIs, movimientos bancarios o scoring).
      parameters:
        - $ref: '#/components/parameters/EmpresaId'
      responses:
        '200':
          description: Lista de periodos disponibles
          content:
            application/json:
              schema:
                type: object
                properties:
                  periodos:
                    type: array
                    items:
                      type: string
              example:
                periodos: ["2024-01", "2023-12", "2023-11"]
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/empresas/{empresa_id}/cierre/{periodo}:
    get:
      operationId: vistaCierre
      tags: [Cierre]
      summary: Vista de cierre mensual
      description: |
        Vista consolidada para el cierre mensual. Responde:
        - Puede cerrar el periodo?
        - Que bloquea el cierre?
        - Que acciones hay pendientes?

        **Criterios para poder cerrar**:
        - Sin riesgos criticos o altos abiertos
        - Conciliacion >= 80%
      parameters:
        - $ref: '#/components/parameters/EmpresaId'
        - name: periodo
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
          description: Periodo fiscal YYYY-MM
          example: "2024-01"
      responses:
        '200':
          description: Estado del cierre mensual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VistaCierre'
              example:
                periodo: "2024-01"
                puede_cerrar: false
                razon_bloqueo: "2 riesgos altos abiertos · conciliacion al 76.5% (minimo 80%)"
                score: 75.0
                bloqueadores: []
                acciones: []
                conciliacion:
                  sin_cfdi: 5
                  sin_movimiento: 2
                  matches_debiles: 3
                  pct_conciliado: 76.5
                  total: 38
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  # ============================================================
  # EMITIDOS / RECIBIDOS
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenido al registrarse o hacer login.
        Incluir en el header: Authorization: Bearer {token}
        Duracion: 8 horas.
  parameters:
    EmpresaId:
      name: empresa_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: ID unico de la empresa (UUID)
      example: "a1b2c3d4-0001-0001-0001-000000000001"
  responses:
    UnauthorizedError:
      description: Token JWT ausente, invalido o expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Se requiere autenticacion"
    ForbiddenError:
      description: El usuario no tiene acceso a este recurso
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Sin acceso a esta empresa"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Descripcion del error
...
    ResumenConciliacion:
      type: object
      properties:
        total:
          type: integer
          description: Total de registros de conciliacion
        exacto:
          type: integer
          description: Matches exactos por monto y RFC
        parcial:
          type: integer
          description: Matches parciales (diferencia tolerable)
        sin_cfdi:
          type: integer
          description: Movimientos bancarios sin CFDI correspondiente
        sin_movimiento:
          type: integer
          description: CFDIs sin movimiento bancario correspondiente
        pct_conciliado:
          type: number
          format: float
          description: Porcentaje de conciliacion (exacto + parcial / total)
    ConciliacionesAccionablesResponse:
      type: object
      properties:
        total:
          type: integer
        pares:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              tipo_match:
                type: string
                enum: [sin_cfdi, parcial]
              monto_movimiento:
                type: number
                format: float
                nullable: true
              monto_cfdi:
                type: number
                format: float
                nullable: true
              diferencia:
                type: number
                format: float
              porcentaje_match:
                type: number
                format: float
              periodo:
                type: string
              movimiento_id:
                type: string
                format: uuid
                nullable: true
              mov_fecha:
                type: string
                format: date
                nullable: true
              concepto:
                type: string
                nullable: true
              mov_monto:
                type: number
                format: float
                nullable: true
              mov_tipo:
                type: string
                enum: [deposito, cargo]
                nullable: true
              rfc_detectado:
                type: string
                nullable: true
    VistaCierre:
      type: object
      properties:
        periodo:
          type: string
        puede_cerrar:
          type: boolean
          description: True si el periodo cumple criterios para cierre
        razon_bloqueo:
          type: string
          nullable: true
          description: Descripcion de lo que impide el cierre
        score:
          type: number
          format: float
          nullable: true
        bloqueadores:
          type: array
          description: Detecciones de riesgo critico/alto que bloquean el cierre
          items:
            type: object
        acciones:
          type: array
          description: Todas las acciones pendientes del periodo
          items:
            type: object
        conciliacion:
          type: object
          properties:
            sin_cfdi:
              type: integer
            sin_movimiento:
              type: integer
            matches_debiles:
              type: integer
            pct_conciliado:
              type: number
              format: float
            total:
              type: integer
...